db_analyze.Teradata {tdplyr}	R Documentation

Analyze the table

Description

This function collects the statistics of the given table.

Usage

## S3 method for class 'Teradata'
db_analyze(con, table, column, ...)

Arguments

con
Required Argument.
Specifies remote data source.

table
Required Argument.
Specifies remote table name.
Types: character

column
Required Argument.
Specifies name of the remote table column(s). Column name is required to collect the statistics in Teradata.
Types: character OR vector of characters

...
Other parameters passed to methods.

Value

Returns 2 on success.

See Also

db_begin, db_commit, db_rollback, sql_table_analyze

Examples

# Get the connection.
con <- td_get_context()$connection

# Load the data.frame 'iris' to "iris" table.
iris2 <- copy_to(con, iris, overwrite = FALSE)

# Analyze the table "iris".
db_analyze(con, "iris", "Sepal.Length")

[Package tdplyr version 17.00.00.02 Index]