print.td_analytics {tdplyr}	R Documentation

Printing tdplyr Analytic Function Objects

Description

Print an object of class "td_analytics".

Usage

## S3 method for class 'td_analytics'
print(x, ...)

Arguments

x
Required Argument.
Specifies a tdplyr analytic function object of class "td_analytics" to print.

...
Optional Arguments.
Specifies further optional arguments to the print method.

Details

This function will automatically be dispatched when the generic function print is invoked to print a specified input object of class "td_analytics".
It calls print on the elements of class "tbl_teradata" from the specified input object.

Examples

# Get remote data source connection.
con <- td_get_context()$connection

# Load example data.
loadExampleData("glm_example", "admissions_train")

# Create object of class "tbl_teradata" to use as input.
admissions_train <- tbl(con, "admissions_train")

# Create the GLM model.
td_glm_out <- td_glm_mle(formula = (admitted ~ stats + masters + gpa + programming),
                         family = "LOGISTIC",
                         linkfunction = "LOGIT",
                         data = admissions_train,
                         weights = "1",
                         threshold = 0.01,
                         maxit = 25,
                         step = FALSE,
                         intercept = TRUE
)

# Print the result which is an object of class "td_analytics".
print(td_glm_out)

[Package tdplyr version 17.00.00.02 Index]