sql_table_analyze.Teradata {tdplyr}	R Documentation

Creates query to analyze the table

Description

This function creates a query to analyze and collect table statistics.

Usage

## S3 method for class 'Teradata'
sql_table_analyze(con, table, column, ...)

Arguments

con
Required Argument.
Specifies remote data source.

table
Required Argument.
Specifies remote table name.
Types: character

column
Required Argument.
Specifies name of the remote table column(s). Column name is required to collect the statistics in Teradata.
Types: character OR vector of characters

...
Other parameters passed to methods.

Value

Returns the SQL query to collect the statistics.

See Also

db_analyze, copy_to

Examples

# Get the connection.
con <- td_get_context()$connection

# Load the data.frame 'iris' to "sql_table_analyze_table" table.
copy_to(con, iris, "sql_table_analyze_table", overwrite = FALSE)

# Get the query to analyze the table "sql_table_analyze_table".
sql_table_analyze(con, "sql_table_analyze_table", "Sepal.Length")

[Package tdplyr version 17.00.00.02 Index]