td_antiselect_sqle {tdplyr}	R Documentation

Antiselect

Description

Antiselect function returns all columns except those specified in the "exclude" argument.
Note: This function is only available when tdplyr is connected to Vantage 1.1 or later versions.

Usage

  td_antiselect_sqle (
    data = NULL,
    exclude = NULL,
    data.order.column = NULL
  )

Arguments

data
Required Argument.
Specifies the name of the input tbl_teradata.

data.order.column
Optional Argument.
Specifies Order By columns for "data".
Values to this argument can be provided as a vector, if multiple columns are used for ordering.
Types: character OR vector of Strings (character)

exclude
Required Argument.
Specifies the names of the columns not to return.
Types: character OR vector of Strings (character)

Value

Function returns an object of class "td_antiselect_sqle" which is a named list containing object of class "tbl_teradata".
Named list member can be referenced directly with the "$" operator using the name: result.

Examples

  
    # Get the current context/connection
    con <- td_get_context()$connection

    # Load example data.
    loadExampleData("antiselect_example", "antiselect_input")

    # Create object(s) of class "tbl_teradata".
    antiselect_input <- tbl(con, "antiselect_input")

    # Example 1 - Exclude five columns as specified in the "exclude" argument
    td_antiselect_sqle_out <- td_antiselect_sqle(data = antiselect_input,
                                                 exclude = c("rowids","orderdate","discount",
                                                             "province", "custsegment")
                                                )
  

[Package tdplyr version 17.00.00.02 Index]