td_copy_files_from_container {tdplyr}	R Documentation

Copy files from container

Description

Function to copy file(s) from container, specified by "container" argument or container indicated by the option 'sandbox.container', to local filesystem.
Notes:

Files are copied to a directory ".tdplyr" directory in user's home directory. The user can get home directory using Sys.getenv("HOME"). If this is not present, then user can find the files in R home directory which can be obtained using Sys.getenv("R_HOME").

User is responsible for cleaning up this directory.

If copying succeeds for one or more files, the list of files for which copying is successful and the location in which files are copied are displayed in console output.

If copying fails for one or more files, then the failed file names and the reasons for failure are displayed in the console output.

If there are duplicate elements in the argument "files.to.copy", then all the duplicated elements are copied only once.

Usage

td_copy_files_from_container(files.to.copy, container = NULL)

Arguments

files.to.copy
Required Argument.
Specifies the file name(s) to be copied from the container.
Note:

Files should be in "/home/tdatuser" inside the container.

Types: character OR vector of Strings

container
Optional Argument.
Specifies the container from which files are to be copied. If not specified, then files are copied from container specified in the option "sandbox.container".
Types: character

Value

TRUE, invisibly.

Examples

# Example 1: Copy files from container started by tdplyr to local client.
td_copy_files_from_container(files.to.copy = "<file1.txt>")

# Example 2: Copy files from container, specified by the argument
#            "container", to local client.
td_copy_files_from_container(files.to.copy = c("<file1.txt>", "<file2.txt>"),
                             container = "<container_id>")

[Package tdplyr version 17.00.00.02 Index]