td_cumulative_mov_avg_mle {tdplyr}	R Documentation

CumulativeMovAvg

Description

The CumulativeMovAvg function computes the cumulative moving average of a value from the beginning of a series.
When computing cumulative moving average, the data are added to the data set in an ordered data stream over time. The objective is to compute the average of all the data at each point in time when new data arrived.

Usage

  td_cumulative_mov_avg_mle (
      data = NULL,
      target.columns = NULL,
      data.sequence.column = NULL,
      data.partition.column = NULL,
      data.order.column = NULL
  )

Arguments

data
Required Argument.
Specifies the input tbl_teradata that contains the columns.

data.partition.column
Required Argument.
Specifies Partition By columns for "data".
Values to this argument can be provided as a vector, if multiple columns are used for partition.
Types: character OR vector of Strings (character)

data.order.column
Required Argument.
Specifies Order By columns for "data".
Values to this argument can be provided as a vector, if multiple columns are used for ordering.
Types: character OR vector of Strings (character)

target.columns
Optional Argument.
Specifies the input column names for which the moving average is to be computed. If you omit this argument, then the function copies every input column to the output tbl_teradata but does not compute moving average.
Types: character OR vector of Strings (character)

data.sequence.column
Optional Argument.
Specifies the vector of column(s) that uniquely identifies each row of the input argument "data". The argument is used to ensure deterministic results for functions which produce results that vary from run to run.
Types: character OR vector of Strings (character)

Value

Function returns an object of class "td_cumulative_mov_avg_mle" which is a named list containing object of class "tbl_teradata".
Named list member can be referenced directly with the "$" operator using name: result.

Examples

  
    # Get the current context/connection
    con <- td_get_context()$connection
    
    # Load example data.
    loadExampleData("cumulativemovavg_example", "ibm_stock")

    # Create object(s) of class "tbl_teradata".
    ibm_stock <- tbl(con, "ibm_stock")

    # Compute the cumulative moving average for "stockprice".
    td_cumulative_mov_avg_out <- td_cumulative_mov_avg_mle(data = ibm_stock,
                                                           data.partition.column = c("name"),
                                                           data.order.column = c("period"),
                                                           target.columns = c("stockprice")
                                                           )
  

[Package tdplyr version 17.00.00.02 Index]