td_execute_script {tdplyr}	R Documentation

Execute user script

Description

The function enables the user to execute script on Teradata Vantage.

Usage

td_execute_script(object, ...)

Arguments

object
Required Argument.
Specifies the ScriptTableOperator object containing the details that are needed for executing the user script.

...
Any additional arguments, if required.

[Package tdplyr version 17.00.00.02 Index]