td_get_context {tdplyr}	R Documentation

Get the current context

Description

Returns a named list of attributes in an existing Teradata context.

Usage

td_get_context()

Value

A named list containing attributes to the current context. NULL if a context does not exist.

See Also

td_create_context , td_set_context

Examples

td_create_context(dsn = "TeradataDSN")
context <- td_get_context()

# To retrieve the connection in the current context, use
td_get_context()$connection

# To retrieve the temorary database in the current context, use
td_get_context()$temp.database

# To retrieve the default database in the current context, use
td_get_context()$database

td_remove_context()

[Package tdplyr version 17.00.00.02 Index]