td_glm_predict_sqle {tdplyr}	R Documentation

GLMPredict

Description

The GLMPredict function uses the model generated by the GLM td_glm_mle function to perform generalized linear model prediction on new input data.

Usage

  td_glm_predict_sqle (
      modeldata = NULL,
      newdata = NULL,
      terms = NULL,
      family = NULL,
      linkfunction = "CANONICAL",
      newdata.order.column = NULL
  )
## S3 method for class 'td_glm_mle'
predict(
      modeldata = NULL,
      newdata = NULL,
      terms = NULL,
      family = NULL,
      linkfunction = "CANONICAL",
      newdata.order.column = NULL)

Arguments

modeldata
Required Argument.
Specifies the model tbl_teradata generated by td_glm_mle.
This argument can accept either a tbl_teradata or an object of "td_glm_mle" class.

newdata
Required Argument.
Specifies the tbl_teradata containing the input data.

newdata.order.column
Optional Argument.
Specifies Order By columns for "newdata".
Values to this argument can be provided as a vector, if multiple columns are used for ordering.
Types: character OR vector of Strings (character)

terms
Optional Argument.
Specifies the names of "newdata" columns to copy to the output tbl_teradata.
Types: character OR vector of Strings (character)

family
Optional Argument.
Specifies the distribution exponential family. The default value is read from "modeldata". If you specify this argument, you must give it the same value that you used for the Family argument of the function when you generated the model table.
Permitted Values: LOGISTIC, BINOMIAL, POISSON, GAUSSIAN, GAMMA, INVERSE_GAUSSIAN, NEGATIVE_BINOMIAL
Types: character

linkfunction
Optional Argument.
The canonical link functions (default link functions) and the link functions that are allowed for each exponential family.
Note: Use the same value that you used for the "linkfunction" argument of the function td_glm_mle when you generated the model.
Default Value: "CANONICAL"
Permitted Values: CANONICAL, IDENTITY, INVERSE, LOG, COMPLEMENTARY_LOG_LOG, SQUARE_ROOT, INVERSE_MU_SQUARED, LOGIT, PROBIT, CAUCHIT
Types: character

Value

Function returns an object of class "td_glm_predict_sqle" which is a named list containing object of class "tbl_teradata".
Named list member can be referenced directly with the "$" operator using name: result.

Examples

  
    # Get the current context/connection
    con <- td_get_context()$connection
    
    # Load example data.
    loadExampleData("glm_example", "admissions_train", "housing_train")
    loadExampleData("glmpredict_example", "admissions_test", "housing_test")
    
    # Create object(s) of class "tbl_teradata".
    admissions_test <- tbl(con, "admissions_test")
    admissions_train <- tbl(con, "admissions_train")
    housing_test <- tbl(con, "housing_test")
    housing_train <- tbl(con, "housing_train")
    
    # Example 1 -
    # First train the data, i.e., create a GLM Model
    td_glm_out <- td_glm_mle(formula = (admitted ~ stats + masters + gpa + programming),
                         family = "LOGISTIC",
                         linkfunction = "LOGIT",
                         data = admissions_train,
                         weights = "1",
                         threshold = 0.01,
                         maxit = 25,
                         step = FALSE,
                         intercept = TRUE
                         )
    
    # Run predict on the output of td_glm_mle() function.
    td_glm_predict_out1 <- td_glm_predict_sqle(modeldata = td_glm_out,
                             newdata = admissions_test,
                             terms = c("id","masters","gpa","stats","programming","admitted"),
                             family = "LOGISTIC",
                             linkfunction = "LOGIT"
                             )
    
    # Example 2 -
    # First train the data, i.e., create a GLM Model
    td_glm_out_hs <- td_glm_mle(formula = (price  ~ recroom  + lotsize  + stories  + garagepl
                                           + gashw + bedrooms  + driveway  + airco  + homestyle
                                           + bathrms  + fullbase + prefarea),
                            family = "GAUSSIAN",
                            linkfunction = "IDENTITY",
                            data = housing_train,
                            weights = "1",
                            threshold = 0.01,
                            maxit = 25,
                            step = FALSE,
                            intercept = TRUE
                            )
    
    # Run predict on the output of td_glm_mle() function by passing coefficients.
    td_glm_predict_out2 <- td_glm_predict_sqle(modeldata = td_glm_out_hs$coefficients,
                             newdata = housing_test,
                             terms = c("sn", "price"),
                             family = "GAUSSIAN",
                             linkfunction = "CANONICAL"
                             )
    
    # Alternatively use S3 predict method to find predictions.
    td_glm_predict_out3 <- predict(td_glm_out_hs,
                             newdata = housing_test,
                             terms = c("sn", "price"),
                             family = "GAUSSIAN",
                             linkfunction = "CANONICAL"
                             )
  

[Package tdplyr version 17.00.00.02 Index]