td_publish_model {tdplyr}	R Documentation

td_publish_model

Description

Function to publish a saved model.
It can be used to update the access level from PRIVATE to PUBLIC or TEAM.
It can also be used to update the status of the model to one of the permitted values.
A model can only be published by its creator.

Usage

td_publish_model(name, grantee = NULL, status = NULL)

Arguments

name
Required Argument.
Specifies the name of the model to be published.
Types: character

grantee
Optional Argument. Must be specified when "status" is not specified.
Specifies a user or role (including PUBLIC) to update the model's access level to.
Types: character

status
Optional Argument. Must be specified when "grantee" is not specified.
Specifies a status to set the the model's status to.
Permitted Values: "ACTIVE", "RETIRED", "CANDIDATE", "PRODUCTION", "IN-DEVELOPMENT"
Types: character

Examples

# Get remote data source connection.
con <- td_get_context()$connection

# Load example data.
loadExampleData("glm_example", "admissions_train")

# Create object of class "tbl_teradata" to use as input.
admissions_train <- tbl(con, "admissions_train")

# Create the GLM model.
td_glm_out1 <- td_glm_mle(formula = (admitted ~ stats + masters + gpa + programming),
                          family = "LOGISTIC",
                          linkfunction = "LOGIT",
                          data = admissions_train,
                          weights = "1",
                          threshold = 0.01,
                          maxit = 25,
                          step = FALSE,
                          intercept = TRUE
)

# Save this generated model.
td_save_model(model = td_glm_out1, name = "glm_model", description = "GLM test")

# Example 1 - Update only the access.
td_publish_model('glm_model', grantee='john')

# Example 2 - Update only the status.
td_publish_model('glm_model', status='Active')

# Example 3 - Update both the access and status.
# This example does not apply to the model we just saved, as we can update
# the access only when the current access is 'Private'.
# We save the model again with another name for this example.
td_save_model(model = td_glm_out1, name = "glm_model_2", description = "GLM test 2")

td_publish_model('glm_model_2', grantee='PUBLIC', status='Production')

[Package tdplyr version 17.00.00.02 Index]