td_scale_summary_mle {tdplyr}	R Documentation

ScaleSummary

Description

The ScaleSummary function takes the ScaleMap (td_scale_map_mle) output (statistics assembled at the vworker level) as input and outputs global statistical information for the entire input data set.

Usage

  td_scale_summary_mle (
    object = NULL,
    object.order.column = NULL
  )

Arguments

object
Required Argument. Specifies the output of td_scale_map_mle function.

object.order.column
Optional Argument.
Specifies Order By columns for "object".
Values to this argument can be provided as a vector, if multiple columns are used for ordering.
Types: character OR vector of Strings (character)

Value

Function returns an object of class "td_scale_summary_mle" which is a named list containing object of class "tbl_teradata".
Named list member can be referenced directly with the "$" operator using name: result.

Examples

  
    # Get the current context/connection
    con <- td_get_context()$connection
    
    # Load example data.
    loadExampleData("scale_example", "scale_stat")

    # Create object(s) of class "tbl_teradata".
    scale_stat_input <- tbl(con, "scale_stat")

    # Example 1 - This function takes the ScaleMap output (scale_stat) as input
    # and outputs global statistical information for the entire input data set.
    td_scale_summary_out <- td_scale_summary_mle(object = scale_stat_input)
  

[Package tdplyr version 17.00.00.02 Index]