td_setup_test_env {tdplyr}	R Documentation

(Deprecated) Setup test environment

Description

The function enables the user to load the sandbox image that is downloaded from Teradata site. This sets up the environment for the user to run the script within docker container using data from csv file, which enables the user to fix script issues outside Teradata Vantage.
Note:

This function is deprecated from tdplyr 17.00.00.01. Teradata recommends using td_setup_sandbox_env() function.

Refer to Teradata Package for R User Guide to get the location of the downloadable docker image.

Usage

td_setup_test_env(docker.image.location = NULL)

Arguments

docker.image.location
Required Argument.
Specifies the location of the downloaded image on user's system.
Types: character

Value

Nothing. Loads the docker image if successful.

See Also

Script, td_test_script

Examples

# Replace "<path_to_docker_image>" with the local path to the downloaded
# docker image file.
docker_image_location <- "<path_to_docker_image>"

# Setup the environment by providing the docker image location.
td_setup_test_env(docker.image.location = docker_image_location)

[Package tdplyr version 17.00.00.02 Index]