td_text_classifier_evaluator_mle {tdplyr}	R Documentation

TextClassifierEvaluator

Description

The TextClassifierEvaluator function evaluates the precision, recall, and F-measure of the trained model output by the TextClassifier (td_text_classifier_mle) function.

Usage

  td_text_classifier_evaluator_mle (
      object = NULL,
      obs.column = NULL,
      predict.column = NULL,
      object.sequence.column = NULL,
      object.order.column = NULL
  )

Arguments

object
Required Argument.
Specifies the tbl_teradata of the trained model output generated by td_text_classifier_mle.
This argument can accept either a tbl_teradata or an object of "td_text_classifier_mle" class.

object.order.column
Optional Argument.
Specifies Order By columns for "object".
Values to this argument can be provided as a vector, if multiple columns are used for ordering.
Types: character OR vector of Strings (character)

obs.column
Required Argument.
Specifies the name of the input column that contains the expected (correct) category.
Types: character

predict.column
Required Argument.
Specifies the name of the input column that contains the predicted category.
Types: character

object.sequence.column
Optional Argument.
Specifies the vector of column(s) that uniquely identifies each row of the input argument "object". The argument is used to ensure deterministic results for functions which produce results that vary from run to run.
Types: character OR vector of Strings (character)

Value

Function returns an object of class "td_text_classifier_evaluator_mle" which is a named list containing object of class "tbl_teradata".
Named list member can be referenced directly with the "$" operator using the name: result.

Examples

  
    # Get the current context/connection
    con <- td_get_context()$connection
    
    # Load example data.
    loadExampleData("text_classifier_trainer_example", "texttrainer_input")
    loadExampleData("text_classifier_example", "textclassifier_input")
    
    # Create object(s) of class "tbl_teradata".
    texttrainer_input <- tbl(con, "texttrainer_input")
    textclassifier_input <- tbl(con, "textclassifier_input")
    
    # The model file 'knn.bin' is generated by td_text_classifier_trainer_mle() function.
    td_text_classifier_trainer_mle(data=texttrainer_input,
                                   text.column='content',
                                   category.column='category',
                                   classifier.type='knn',
                                   model.file='knn.bin',
                                   classifier.parameters='compress:0.9',
                                   nlp.parameters=c('useStem:true','stopwordsFile:stopwords.txt'),
                                   feature.selection='DF:[0.1:0.99]',
                                   data.sequence.column='id'
                                   )
    
     
    # The generated model file is used by td_text_classifier_mle() function to classify the 
    # input text.
    td_text_classifier_out <- td_text_classifier_mle(newdata=textclassifier_input,
                                                     newdata.order.column='id',
                                                     text.column='content',
                                                     accumulate=c('id','category'),
                                                     model.file='knn.bin',
                                                     newdata.sequence.column='id'
                                                     )
    
    # Example 1 - This example evaluates the precision, recall, and F-measure
    # of the model output generated by td_text_classifier_mle() function.
    text_classifier_eval_out <- td_text_classifier_evaluator_mle(object = td_text_classifier_out,
                                                                 obs.column = 'category',
                                                                 predict.column = 'out_category',
                                                                 object.sequence.column = 'id',
                                                                 object.order.column = 'id'
                                                                 )
  

[Package tdplyr version 17.00.00.02 Index]