tdplyr-package {tdplyr}	R Documentation

tdplyr: dplyr for Teradata Vantage

Description

tdplyr is focused on integrating and scaling dplyr with Teradata Vantage to solve challenges in data analysis.

Details

The main goals are to:

Enable data scientists to work effectively with the data in Teradata Vantage.

Provide a growing list of analytic functions to leverage in-database analytics.

To learn more about tdplyr, start with the vignettes:

browseVignettes(package = "tdplyr")

Use the help command to view the list of exported functions:

help(package = "tdplyr")

To learn more about dplyr, start with the vignettes:

browseVignettes(package = "dplyr")

Package options

td.db_desc.output

Sets the Database output when printing tbls

Package options for dplyr related packages

Options from the dplyr package

dplyr.strict_sql

Forces an error if dplyr doesn't know how to map a function when translating SQL if TRUE. FALSE by default.

Options from the tibble package

tibble.print_min

Sets minimum amount of rows to print by default

Glossary of Terms

Teradata tbl object: Refers to the object created using the tbl() function of the "dplyr" package, when the function is invoked with a Teradata connection. Teradata tbl is an object of class "tbl_teradata". For example, to create a Teradata tbl object from an existing table "TABLE1" with a connection "con", use the command: dplyr::tbl(con, "TABLE1").

[Package tdplyr version 17.00.00.02 Index]